Split prioritization into scheme and header sections
I added a bit of text to introduce the 'Urgency' prioritization scheme, but besides that this is just moving text around.  I think the header section now needs more explanation, but that can be added in a different PR.

I'm happy to pick a different name from 'Urgency' as well, but it seemed like the most obvious choice.